Fix type errors and a few warnings for Visual Studio 2010 64-bit build. We're still far from 64b statically safe, mostly around SkReader32.h and the address-alignment code in GrTypes.h. Original code provided by jianliang79.
http://code.google.com/p/skia/issues/detail?id=601



git-svn-id: http://skia.googlecode.com/svn/trunk@3994 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ gm/shadows.cpp
@@ -95,7 +95,7 @@ protected:
 
         paint.setLooper(shadowLoopers[i]);
 
-        canvas->translate(SkIntToScalar(i*40), SkIntToScalar(0));
+        canvas->translate(SkIntToScalar((unsigned int)i*40), SkIntToScalar(0));
         setup(&paint, gRec[0].fColor, gRec[0].fStrokeWidth);
         canvas->drawRect(fRect, paint);
